feat: add support for Unity 6000.0 or later
close #141

+++ Packages/src/Editor/Coffee.UpmGitExtension/GitPackageDataBase.cs
@@ -7,9 +7,9 @@ using UnityEditor;
 using UnityEditor.PackageManager;
 using UnityEditorInternal;
 using UnityEngine;
+
 #if UNITY_2021_1_OR_NEWER
 using UnityEditor.PackageManager.UI.Internal;
-
 #else
 using UnityEditor.PackageManager.UI;
 #endif
@@ -26,7 +26,12 @@ namespace Coffee.UpmGitExtension
         public string id;
         public string url;
         public int hash;
+        
+#if UNITY_6000_0_OR_NEWER
+        public UpmPackageVersion[] versions;
+#else
         public UpmPackageVersionEx[] versions;
+#endif
 
         void ISerializationCallbackReceiver.OnBeforeSerialize()
         {
@@ -34,9 +39,13 @@ namespace Coffee.UpmGitExtension
 
         void ISerializationCallbackReceiver.OnAfterDeserialize()
         {
+#if UNITY_6000_0_OR_NEWER
+            versions = versions.ToArray();
+#else
             versions = versions
                 .Where(v => v.isValid)
                 .ToArray();
+#endif
         }
 
         public override int GetHashCode()
@@ -49,18 +58,19 @@ namespace Coffee.UpmGitExtension
             return (obj as FetchResult)?.hash == hash;
         }
     }
-
-    /// <summary>
-    /// Database of packages installed via Git
-    /// </summary>
-    //[FilePath("GitPackageDatabase.asset", FilePathAttribute.Location.PreferencesFolder)]
+    
     internal class GitPackageDatabase : ScriptableSingleton<GitPackageDatabase>
     {
-        //################################
-        // Public Members.
-        //################################
-        // public static event Action OnChangedPackages;
-
+        private static string _workingDirectory => InternalEditorUtility.unityPreferencesFolder + "/GitPackageDatabase";
+        private static string _serializeVersion => "2.0.2";
+        private static string _resultsDir => _workingDirectory + "/Results-" + _serializeVersion;
+        private static FileSystemWatcher _watcher;
+        private static bool _isPaused;
+        private static readonly HashSet<FetchResult> _resultCaches = new HashSet<FetchResult>();
+        private static PackageManagerProjectSettings _settings => ScriptableSingleton<PackageManagerProjectSettings>.instance;
+        internal static UpmClient _upmClient => ScriptableSingleton<ServicesContainer>.instance.Resolve<UpmClient>();
+        internal static PackageDatabase _packageDatabase => ScriptableSingleton<ServicesContainer>.instance.Resolve<PackageDatabase>();
+        
         public static void Install(string packageId)
         {
             _upmClient.AddByUrl(packageId);
@@ -85,10 +95,7 @@ namespace Coffee.UpmGitExtension
             return GetUpmPackages()
                 .Where(p => p.GetInstalledVersion()?.HasTag(PackageTag.Git) == true);
         }
-
-        /// <summary>
-        /// Fetch the available git package versions.
-        /// </summary>
+        
         public static void Fetch(string url, Action<int> callback = null)
         {
             const string kFetchPackagesJs = "Packages/com.coffee.upm-git-extension/Editor/Commands/fetch-packages.js";
@@ -112,7 +119,17 @@ namespace Coffee.UpmGitExtension
         {
             return _packageDatabase.GetPackage(packageName);
         }
+        
+#if UNITY_6000_0_OR_NEWER
+        internal static List<UpmPackageVersion> GetPackageVersion(string packageName, string versionUniqueId)
+        {
+            var result = _resultCaches
+                .SelectMany(r => r.versions)
+                .Where(v => v.name == packageName).ToList();
 
+            return result;
+        }
+#else
         internal static IPackageVersion GetPackageVersion(string packageUniqueId, string versionUniqueId)
         {
             IPackage package;
@@ -120,10 +137,8 @@ namespace Coffee.UpmGitExtension
             _packageDatabase.GetPackageAndVersion(packageUniqueId, versionUniqueId, out package, out version);
             return version;
         }
-
-        /// <summary>
-        /// Update available versions for git packages.
-        /// </summary>
+#endif
+        
         public static void Fetch()
         {
             GetInstalledGitPackages()
@@ -149,17 +164,20 @@ namespace Coffee.UpmGitExtension
             {
                 foreach (var dir in Directory.GetDirectories(_workingDirectory))
                 {
-                    Directory.Delete(dir, true);
+                    if (Directory.Exists(dir))
+                    {
+                        Directory.Delete(dir, true);
+                    }
                 }
             }
 
             Debug.Log("[GitPackageDatabase] Clear Cache");
             WatchResultJson();
         }
-
+        
         public static void ResetCacheTime()
         {
-            isPaused = true;
+            _isPaused = true;
             var resultDir = Path.GetFullPath(_resultsDir);
             if (Directory.Exists(resultDir))
             {
@@ -169,42 +187,26 @@ namespace Coffee.UpmGitExtension
                 }
             }
 
-            isPaused = false;
+            _isPaused = false;
         }
+        
+#if UNITY_6000_0_OR_NEWER
+        public static IEnumerable<UpmPackageVersion> GetAvailablePackageVersions(string repoUrl = null)
+        {
+            var result = _resultCaches
+                .SelectMany(r => r.versions)
+                .Where(v => (string.IsNullOrEmpty(repoUrl) || v.uniqueId.Contains(repoUrl)));
 
+            return result;
+        }
+#else
         public static IEnumerable<UpmPackageVersionEx> GetAvailablePackageVersions(string repoUrl = null)
         {
-            return _resultCaches
-                .SelectMany(r => r.versions)
+            return _resultCaches.SelectMany(r => r.versions)
                 .Where(v => v.isValid && (string.IsNullOrEmpty(repoUrl) || v.uniqueId.Contains(repoUrl)));
         }
-
-        //################################
-        // Private Members.
-        //################################
-        private static string _workingDirectory => InternalEditorUtility.unityPreferencesFolder + "/GitPackageDatabase";
-        private static string _serializeVersion => "2.0.2";
-        private static string _resultsDir => _workingDirectory + "/Results-" + _serializeVersion;
-        private static FileSystemWatcher _watcher;
-        private static bool isPaused;
-        private static readonly HashSet<FetchResult> _resultCaches = new HashSet<FetchResult>();
-        private static PackageManagerProjectSettings _settings =>
-            ScriptableSingleton<PackageManagerProjectSettings>.instance;
-#if UNITY_2020_2_OR_NEWER
-        internal static UpmClient _upmClient => ScriptableSingleton<ServicesContainer>.instance.Resolve<UpmClient>();
-        internal static PackageDatabase _packageDatabase =>
-            ScriptableSingleton<ServicesContainer>.instance.Resolve<PackageDatabase>();
-#else
-        internal static IUpmClient _upmClient => UpmClient.instance;
-        internal static IPackageDatabase _packageDatabase => PackageDatabase.instance;
 #endif
-
-#if UNITY_2021_1_OR_NEWER
-        private static bool _enablePreReleasePackages => _settings.enablePreReleasePackages;
-#else
-        private static bool _enablePreReleasePackages => _settings.enablePreviewPackages;
-#endif
-
+        
         public static void RequestUpdateGitPackageVersions()
         {
             EditorApplication.delayCall -= UpdateGitPackageVersions;
@@ -224,29 +226,35 @@ namespace Coffee.UpmGitExtension
                 .Select(versions =>
                 {
                     var isInstalled = installedIds.Contains(versions.Key);
-                    if (!isInstalled) return null;
-
+                    if (!isInstalled)
                     {
-                        // Git mode: Register all installable package versions.
-                        var upmPackage = _packageDatabase.GetPackage(versions.Key) as UpmPackage;
-                        var installedVersion = upmPackage.versions.installed as UpmPackageVersion;
-                        if (installedVersion.GetPackageInfo().source != PackageSource.Git)
-                        {
-                            return upmPackage;
-                        }
-
-                        // Unlock.
-                        installedVersion.UnlockVersion();
-
-                        var newVersions = new[] { new UpmPackageVersionEx(installedVersion) }
-                            .Concat(versions.Where(v => v.uniqueId != installedVersion.uniqueId))
-                            .OrderBy(v => v.semVersion)
-                            .ThenBy(v => v.isInstalled)
-                            .ToArray();
-                        upmPackage = upmPackage.UpdateVersionsSafety(newVersions);
-
+                        return null;
+                    }
+                    
+                    // Git mode: Register all installable package versions.
+                    
+                    var upmPackage = _packageDatabase.GetPackage(versions.Key) as UpmPackage;
+                    var installedVersion = upmPackage?.versions.installed as UpmPackageVersion;
+                    if (installedVersion.GetPackageInfo().source != PackageSource.Git)
+                    {
                         return upmPackage;
                     }
+                    
+                    // Unlock.
+                    installedVersion.UnlockVersion();
+                    
+#if UNITY_6000_0_OR_NEWER
+                    upmPackage = upmPackage.UpdateVersionsSafety();
+#else
+                    var newVersions = new[] { new UpmPackageVersionEx(installedVersion) }
+                        .Concat(versions.Where(v => v.uniqueId != installedVersion.uniqueId))
+                        .OrderBy(v => v.semVersion)
+                        .ThenBy(v => v.isInstalled)
+                        .ToArray();
+                    upmPackage = upmPackage.UpdateVersionsSafety(newVersions);
+#endif
+
+                    return upmPackage;
                 })
                 .Where(p => p != null);
 
@@ -272,7 +280,7 @@ namespace Coffee.UpmGitExtension
 
         private static void OnResultFileCreated(string file)
         {
-            if (isPaused || string.IsNullOrEmpty(file) || Path.GetExtension(file) != ".json" || !File.Exists(file))
+            if (_isPaused || string.IsNullOrEmpty(file) || Path.GetExtension(file) != ".json" || !File.Exists(file))
             {
                 return;
             }
@@ -324,6 +332,38 @@ namespace Coffee.UpmGitExtension
             _watcher.Created += (s, e) => EditorApplication.delayCall += () => OnResultFileCreated(e.FullPath);
 
             _upmClient.onAddOperation += op => op.onOperationFinalized += _ => RequestUpdateGitPackageVersions();
+        }
+        
+        public static string GetShortPackageId(UpmPackageVersion self)
+        {
+            var semver = self.versionString;
+            var revision = ExtractGitRevision(self.uniqueId);
+
+            return !string.IsNullOrEmpty(revision) && !revision.Contains(semver)
+                ? $"{self.name}/{revision} ({semver})"
+                : $"{self.name}/{semver}";
+        }
+        
+        public static string GetShortVersion(UpmPackageVersion self)
+        {
+            var semver = self.versionString;
+            var revision = ExtractGitRevision(self.uniqueId);
+
+            return !string.IsNullOrEmpty(revision) && !revision.Contains(semver)
+                ? $"{revision} ({semver})"
+                : $"{semver}";
+        }
+        
+        private static string ExtractGitRevision(string uniqueId)
+        {
+            if (string.IsNullOrEmpty(uniqueId))
+                return null;
+
+            var hashIndex = uniqueId.LastIndexOf('#');
+            if (hashIndex < 0 || hashIndex == uniqueId.Length - 1)
+                return null;
+
+            return uniqueId.Substring(hashIndex + 1);
         }
     }
 }
